# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PicoClaw Serverless API
  version: 1.0.0
  description: |
    HTTP API for PicoClaw (serverless-oriented runtime).

    Notes:
    - All endpoints except `/health` require Bearer token authentication when `API_TOKEN` is set. When `API_TOKEN` is unset, authentication is disabled (auth-free mode for local/trusted-network deployments).
    - `POST /chat` supports both JSON response mode and SSE streaming mode.
    - Memory and conversation history are persisted via mounted storage and reused across separate requests.
    - Scheduled tasks are externally triggered by calling `/task/check`.
    - `POST /control/stop` can be used to request graceful save-and-exit.
    - Every response includes an `X-Request-ID` header for log correlation. If the caller sends an `X-Request-ID` header, the same value is echoed back; otherwise a new `req-<UUID>` is generated.
    - Every response includes `X-Build-Version` and `X-Build-Commit` headers for identifying the running build.
servers:
  - url: http://localhost:9000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /chat:
    post:
      tags:
        - Chat
      summary: Start or continue a conversation
      description: |
        Sends a user message to the agent.

        Behavior:
        - Without `conversation_id`: creates a new conversation.
        - With `conversation_id`: resumes an existing conversation.
        - With `stream=true`: returns `text/event-stream` (SSE).
      operationId: postChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              newConversation:
                value:
                  message: 请只回答数字：1+1 等于几？
                  sender: verify
                  sender_name: Verifier
                  stream: false
              continueConversation:
                value:
                  message: 上一轮你回答了什么？只回答内容。
                  conversation_id: conv-0df6aee0-86e1-45ac-b33b-c13086b72c36
                  sender: verify
                  sender_name: Verifier
                  stream: false
              streamMode:
                value:
                  message: 逐步思考后给结论
                  sender: verify
                  sender_name: Verifier
                  stream: true
                  max_execution_ms: 120000
              withMcpServer:
                value:
                  message: 请帮我分析一下最近一周的支出情况
                  mcp_servers:
                    finance:
                      type: http
                      url: http://example.com/mcp-server
      responses:
        '200':
          description: Chat processed successfully (JSON or SSE stream)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                success:
                  value:
                    status: success
                    conversation_id: conv-0df6aee0-86e1-45ac-b33b-c13086b72c36
                    message_id: msg-3aaf5d3f-a17b-4e8e-bec3-a33df4c9af86
                    result: '2'
                    model: claude-sonnet-4-6
                    session_id: 3e49773a-b4a6-4c53-8cd9-b3f96c5377e3
                    duration_ms: 6701
                    error: null
                    outbound_messages: []
                    session_end_marker: '[[PICOCLAW_SESSION_END]]'
                    session_end_marker_detected: false
                timeout:
                  value:
                    status: timeout
                    conversation_id: conv-aaa
                    message_id: msg-bbb
                    result: 已生成部分结果
                    session_id: session-xxx
                    duration_ms: 120000
                    error: Execution aborted after 120000ms. Use conversation_id to continue.
                    outbound_messages: []
                    session_end_marker: '[[PICOCLAW_SESSION_END]]'
                    session_end_marker_detected: false
            text/event-stream:
              schema:
                type: string
              examples:
                sse:
                  value: |
                    event: start
                    data: {"conversation_id":"conv-123","message_id":"msg-123"}

                    event: thinking
                    data: {"text":"Let me reason about this..."}

                    event: tool_use
                    data: {"tool":"WebSearch","input":{"query":"example"}}

                    event: chunk
                    data: {"text":"partial output"}

                    event: done
                    data: {"status":"success","conversation_id":"conv-123","message_id":"msg-123","result":"final result","session_id":"session-123","duration_ms":3210,"error":null,"outbound_messages":[],"session_end_marker":"[[PICOCLAW_SESSION_END]]","session_end_marker_detected":true}
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                missingMessage:
                  value:
                    error: message is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation ID not found when resuming
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                notFound:
                  value:
                    error: 'conversation_id not found: conv-missing'
        '409':
          description: Conversation is currently processing another request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationBusyError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatInternalErrorResponse'
    get:
      tags:
        - Chat
      summary: List all conversations
      operationId: listConversations
      responses:
        '200':
          description: Conversation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/{conversation_id}:
    get:
      tags:
        - Chat
      summary: Get conversation state
      operationId: getConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Conversation state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationStatusResponse'
        '400':
          description: Missing or invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Chat
      summary: Delete a conversation
      description: |
        Deletes a conversation and all related data (messages, outbound messages, tasks).
        Returns 409 if the conversation is currently running.
      operationId: deleteConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '204':
          description: Conversation deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conversation is currently running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationBusyError'
  /chat/{conversation_id}/messages:
    get:
      tags:
        - Chat
      summary: Get conversation messages
      operationId: getConversationMessages
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationMessagesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_TOKEN
  parameters:
    ConversationId:
      name: conversation_id
      in: path
      required: true
      description: Conversation ID
      schema:
        type: string
        minLength: 1
...
  headers:
    X-Request-ID:
      description: |
        Request tracking identifier for log correlation.
        Echoes the caller-provided `X-Request-ID` header, or a generated `req-<UUID>` if none was sent.
        Present on every response, including error and unauthenticated responses.
      schema:
        type: string
        example: req-a1b2c3d4-e5f6-7890-abcd-ef1234567890
    X-Build-Version:
      description: |
        Application version from `package.json`, injected at Docker build time.
        Present on every response.
      schema:
        type: string
        example: 1.2.16
    X-Build-Commit:
      description: |
        Short git commit hash of the build, injected at Docker build time.
        Present on every response.
      schema:
        type: string
        example: abc1234
  responses:
    Unauthorized:
      description: Unauthorized (missing or invalid bearer token)
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
        X-Build-Version:
          $ref: '#/components/headers/X-Build-Version'
        X-Build-Commit:
          $ref: '#/components/headers/X-Build-Commit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorized:
              value:
                error: Unauthorized
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
...
    ChatRequest:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
        conversation_id:
          type: string
          description: Existing conversation ID for multi-turn continuation.
        sender:
          type: string
          description: Sender identifier.
          default: user
        sender_name:
          type: string
          description: Sender display name.
        stream:
          type: boolean
          default: false
          description: Return SSE stream when true.
        max_execution_ms:
          type: integer
          minimum: 1
          description: |
            Request-level timeout in milliseconds.
            Effective timeout is min(max_execution_ms, env MAX_EXECUTION_MS).
        thinking:
          type: boolean
          default: false
          description: |
            Enable extended thinking. When true, the model produces thinking
            content before its final response. Requires stream=true to see
            thinking events in SSE output.
        max_thinking_tokens:
          type: integer
          minimum: 1
          description: |
            Maximum tokens for extended thinking. Only used when thinking=true.
            Default: 10000.
        show_tool_use:
          type: boolean
          default: false
          description: |
            Stream tool invocation events. When true and stream=true, SSE
            output includes 'tool_use' events showing which tools the agent
            calls and their inputs.
        model:
          type: string
          description: |
            Model override for this request. Omit to use the CLI default
            (recommended — avoids locking to a specific model version).
            Accepts full model IDs (e.g. `claude-opus-4-6`) or short names
            (`opus`, `sonnet`, `haiku`). Precedence: per-request > `CLAUDE_MODEL`
            env var > CLI default.
        mcp_servers:
          type: object
          description: |
            Per-request MCP servers merged with the built-in picoclaw server.
            Each key is a server name; the value defines the transport config.
            Supports stdio, SSE, and HTTP transports (maps to SDK McpServerConfig).
          additionalProperties:
            oneOf:
              - type: object
                required:
                  - url
                properties:
                  type:
                    type: string
                    enum:
                      - http
                      - sse
                  url:
                    type: string
                    format: uri
                  headers:
                    type: object
                    additionalProperties:
                      type: string
              - type: object
                required:
                  - command
                properties:
                  type:
                    type: string
                    enum:
                      - stdio
                  command:
                    type: string
                  args:
                    type: array
                    items:
                      type: string
                  env:
                    type: object
                    additionalProperties:
                      type: string
    OutboundMessage:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - text
        - sender
        - created_at
      properties:
        id:
          type: integer
        conversation_id:
          type: string
        text:
          type: string
        sender:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ChatResponse:
      type: object
      required:
        - status
        - conversation_id
        - message_id
        - result
        - duration_ms
        - outbound_messages
        - session_end_marker
        - session_end_marker_detected
      properties:
        status:
          type: string
          enum:
            - success
            - timeout
            - error
        conversation_id:
          type: string
        message_id:
          type: string
        result:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
          description: Actual model used for this request (from SDK init).
        session_id:
          type: string
          nullable: true
        duration_ms:
          type: integer
          minimum: 0
        error:
          type: string
          nullable: true
        outbound_messages:
          type: array
          items:
            $ref: '#/components/schemas/OutboundMessage'
        session_end_marker:
          type: string
          description: Marker string that indicates the assistant declared session completion.
        session_end_marker_detected:
          type: boolean
          description: True when the marker is found in result or outbound messages.
        usage:
          type: object
          nullable: true
          description: Token usage and cost metrics from the agent execution (present when SDK provides data).
          properties:
            inputTokens:
              type: integer
              minimum: 0
            outputTokens:
              type: integer
              minimum: 0
            totalCostUsd:
              type: number
              minimum: 0
            numTurns:
              type: integer
              minimum: 0
            durationApiMs:
              type: integer
              minimum: 0
        warnings:
          type: array
          items:
            type: string
          description: MCP server validation warnings (reserved names rejected, invalid configs skipped). Only present when issues are detected.
    ChatInternalErrorResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - conversation_id
        - message_id
        - result
        - error
      properties:
        status:
          type: string
          enum:
            - error
        conversation_id:
          type: string
        message_id:
          type: string
        result:
          type: string
          nullable: true
        error:
          type: string
    ConversationStatusResponse:
      type: object
      additionalProperties: false
      required:
        - conversation_id
        - message_count
        - last_activity
        - status
      properties:
        conversation_id:
          type: string
        session_id:
          type: string
          nullable: true
        message_count:
          type: integer
          minimum: 0
        last_activity:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - idle
            - running
...
    ConversationListResponse:
      type: object
      additionalProperties: false
      required:
        - conversations
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationStatusResponse'
    ConversationMessage:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - role
        - content
        - created_at
      properties:
        id:
          type: string
        conversation_id:
          type: string
        role:
          type: string
          enum:
            - user
            - assistant
            - system
        sender:
          type: string
          nullable: true
        sender_name:
          type: string
          nullable: true
        content:
          type: string
        created_at:
          type: string
          format: date-time
    ConversationMessagesResponse:
      type: object
      additionalProperties: false
      required:
        - conversation_id
        - messages
      properties:
        conversation_id:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMessage'
    ConversationBusyError:
      type: object
      additionalProperties: false
      required:
        - error
        - conversation_id
      properties:
        error:
          type: string
        conversation_id:
          type: string
...